move web code to web.py
